fix: revamped approach to use postbuild script
---

After a few more builds I have realized that the ordering was just 'luck' for the few times I have executed it previously. Therefore I have adjusted it with a small custom script. There are a few questions regarding the .d.ts files' overshadowing and how to better influence the creation of the index.d.ts file in the dist folder (mentioned in the readme).